Flightmanagementreport.pdf
This project, created as part of the "Data Structures and Applications" course at Malnad College of Engineering, is a simulation of a Flight Information Display System (FIDS). It demonstrates the use of data structures to develop a functional system for managing and querying flight information. Key features include user account management, direct and indirect flight suggestions, and the ability to find the cheapest routes between cities. The project emphasizes user-friendliness and efficient data handling using structured programming.